% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{class}
\name{LabelEncoder-class}
\alias{LabelEncoder-class}
\title{An S4 class to represent a LabelEncoder.}
\description{
An S4 class to represent a LabelEncoder.
}
\section{Slots}{

\describe{
\item{\code{type}}{A character to denote the input type, either character, factor or numeric}

\item{\code{mapping}}{A data.frame to store the mapping table}
}}

