% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{class}
\name{LabelEncoder.Character-class}
\alias{LabelEncoder.Character-class}
\title{An S4 class to represent a LabelEncoder with character input.}
\description{
An S4 class to represent a LabelEncoder with character input.
}
\section{Slots}{

\describe{
\item{\code{classes}}{A character vector to store the unique values of classes}
}}

