% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{class}
\name{LabelEncoder.Factor-class}
\alias{LabelEncoder.Factor-class}
\title{An S4 class to represent a LabelEncoder with factor input.}
\description{
An S4 class to represent a LabelEncoder with factor input.
}
\section{Slots}{

\describe{
\item{\code{classes}}{A factor vector to store the unique values of classes}
}}

