% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{class}
\name{LabelEncoder.Numeric-class}
\alias{LabelEncoder.Numeric-class}
\title{An S4 class to represent a LabelEncoder with numeric input.}
\description{
An S4 class to represent a LabelEncoder with numeric input.
}
\section{Slots}{

\describe{
\item{\code{classes}}{A numeric vector to store the unique values of classes}
}}

