% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{class}
\name{OneHotEncoder-class}
\alias{OneHotEncoder-class}
\title{An S4 class to represent a OneHotEncoder}
\description{
An S4 class to represent a OneHotEncoder
}
\section{Slots}{

\describe{
\item{\code{n_columns}}{An integer value to store the number of columns of input data}

\item{\code{n_values}}{A numeric vector to store the number of unique values in each column of input data}

\item{\code{column_encoders}}{A list that stores the LabelEncoder for each column of input data}
}}

